+++ src/tiddlywiki-codemirror-6/modules/extensions/iconpatch.ts
@@ -1,0 +1,8 @@
+import { EditorView } from '@codemirror/view';
+
+/** @see-also https://codemirror.net/examples/styling */
+export default EditorView.theme({
+  '.cm-completionIcon': {
+    width: `unset !important`
+  }
+});