+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 <h2>V3</h2>
 
+* `3.8.5`: update `/date` snippets, and add `/username` example for tiddlywiki-cmp plugin
+* `3.8.3`: supprot system snippets for prefix $:/snippets/
+* `3.8.2`: remove $ trigger for fields
+* `3.8.1`: add tree view for plugin(2025-01-25)
 * `3.8.0`: expose cm6 instance for dev
 * `3.7.0`: better newline extension for current line
 * `3.7.0`: `ctrl+click` to double click for tiddler navigation